+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.21.7 (2023-09-27)
+
 - changed: Update getInOutTokenAddresses to use EdgeTokens so it can be used for any token
 - changed: Block VELO trading from providers that rely on `currencyCode`
 - changed: Add `fantom` pluginId check to spookySwap and tombSwap